Hacked in a power on / off detection
The receiver receives a message when the headset turns on / off / connects / disconnects.
Since there is no reason to keep the sinks open if the headsets are off, they now close.
This has the benefit of making the default output device switch to another device (e.g. speakers) instead of having to manually do that.
This assumes the headset is off when the process starts. If on, turning off and on should fix.
It may also be in this script's power to disable the "main" headset sink.
This implementation isn't very expandable. May switch to something more like signals or the already implement on_open and on_close methods.